fix: refactor wrap-external to load-assets
