+++ packages/ice-plugin-load-assets/CHANGELOG.md
@@ -1,9 +1,5 @@
 # Changelog
 
-## 0.1.1
-
-- [refactor] refactor function loadUrls
-
 ## 0.1.0
 
 - [feat] 完成初始版本基础功能
