+++ packages/ice-plugin-load-assets/getLoadScriptsCode.js
